🔨 Refactor: MyRoomCanvas 리팩토링 및 파일 분리
