+++ src/app/myroom/[id]/_component/FurnitureInfoModal/FurnitureInfoModal.tsx
@@ -8,8 +8,8 @@ import {
   DialogHeader,
   DialogTitle,
 } from "@/components/ui/dialog";
+import { FurnitureInfoModalProps } from "@/model/Funiture";
 
-import { FurnitureInfoModalProps } from "../../_model/Funiture";
 import Style from "./FurnitureInfoModal.style";
 
 const FurnitureInfoModal: React.FC<FurnitureInfoModalProps> = ({
